Implement more authorization

+++ services/CharacterService/CharacterService.cs
@@ -1,11 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using dotnetcore.Data;
 using dotnetcore.DTOs.Character;
 using dotnetcore.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace dotnetcore.services.CharacterService
@@ -14,38 +16,43 @@ namespace dotnetcore.services.CharacterService
     {
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
+            _httpContextAccessor = httpContextAccessor;
             _mapper = mapper;
             _context = context;
         }
+
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirst(ClaimTypes.NameIdentifier).Value);
         
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
-            ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
+            ServiceResponse<GetCharacterDto> serviceResponse = new();
             List<Character> characters = await _context.Characters.ToListAsync();
-            serviceResponse.Data = _mapper.Map<GetCharacterDto>(characters.FirstOrDefault(c => c.Id == id));
+            serviceResponse.Data = _mapper.Map<GetCharacterDto>(characters.FirstOrDefault(c => c.Id == id && c.User.Id == GetUserId()));
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetCharacters(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
-            ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            List<Character> characters = await _context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+            ServiceResponse<List<GetCharacterDto>> serviceResponse = new();
+            List<Character> characters = await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
             serviceResponse.Data = _mapper.Map<List<GetCharacterDto>>(characters);
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> PostCharacter(AddCharacterDto newCharacter)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
-            ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
+            ServiceResponse<List<GetCharacterDto>> serviceResponse = new();
             Character character = _mapper.Map<Character>(newCharacter);
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
             
             await _context.Characters.AddAsync(character);
             await _context.SaveChangesAsync();
             
-            serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+            serviceResponse.Data = _context.Characters.Where(c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             return serviceResponse;
         }
 
@@ -77,17 +84,32 @@ namespace dotnetcore.services.CharacterService
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<GetCharacterDto>> DeleteCharacter(int id)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> DeleteCharacter(int id)
         {
-            List<Character> characters = await _context.Characters.ToListAsync();
-            Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-            _context.Characters.Remove(character);
-            await _context.SaveChangesAsync();
+            ServiceResponse<List<GetCharacterDto>> serviceResponse = new();
 
-            ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>
+            try
             {
-                Data = _mapper.Map<GetCharacterDto>(character)
-            };
+                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.Id == GetUserId());
+
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    serviceResponse.Data = _context.Characters.Where(c => c.User.Id == GetUserId())
+                        .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                }
+                else
+                {
+                    serviceResponse.IsSuccess = false;
+                    serviceResponse.Message = "Character not found.";
+                }
+            }
+            catch (Exception ex)
+            {
+                serviceResponse.IsSuccess = false;
+                serviceResponse.Message = ex.Message;
+            }
 
             return serviceResponse;
         }
